refactor(quant): move modelopt_quant into modelopt package
- Move implementation to layers/quantization/modelopt/impl.py
- Re-export public API from layers/quantization/modelopt/__init__.py
- Update imports in quantization registry, weight_utils, fused_moe, tests

No functional changes.

Made-with: Cursor

refactor init
