+++ python/sglang/srt/layers/quantization/modelopt/__init__.py
@@ -1,0 +1,34 @@
+# SPDX-License-Identifier: Apache-2.0
+"""ModelOpt quantization package (layout mirrors ``compressed_tensors``)."""
+
+from __future__ import annotations
+
+from sglang.srt.layers.quantization.modelopt.modelopt import ModelOptQuantConfig
+from sglang.srt.layers.quantization.modelopt.schemes import (
+    ModelOptFp4Config,
+    ModelOptFp4LinearMethod,
+    ModelOptFp8Config,
+    ModelOptFp8KVCacheMethod,
+    ModelOptFp8LinearMethod,
+    ModelOptFp8MoEMethod,
+    ModelOptMixedPrecisionConfig,
+    ModelOptNvFp4FusedMoEMethod,
+    enable_flashinfer_fp4_gemm,
+    fp4_gemm,
+    fp4_quantize,
+)
+
+__all__ = [
+    "ModelOptFp4Config",
+    "ModelOptFp4LinearMethod",
+    "ModelOptFp8Config",
+    "ModelOptFp8KVCacheMethod",
+    "ModelOptFp8LinearMethod",
+    "ModelOptFp8MoEMethod",
+    "ModelOptMixedPrecisionConfig",
+    "ModelOptNvFp4FusedMoEMethod",
+    "ModelOptQuantConfig",
+    "enable_flashinfer_fp4_gemm",
+    "fp4_gemm",
+    "fp4_quantize",
+]
