+++ python/sglang/srt/layers/quantization/modelopt/schemes/__init__.py
@@ -1,0 +1,42 @@
+# SPDX-License-Identifier: Apache-2.0
+
+# FP4 first (shared MoE symbols for FP8); then FP8; then mixed precision.
+from sglang.srt.layers.quantization.modelopt.schemes.modelopt_fp4 import (
+    ACT_STR_TO_TYPE_MAP,
+    ACTIVATION_SCHEMES,
+    MOE_NVFP4_DISPATCH,
+    ActivationType,
+    ModelOptFp4Config,
+    ModelOptFp4LinearMethod,
+    ModelOptNvFp4FusedMoEMethod,
+    enable_flashinfer_fp4_gemm,
+    fp4_gemm,
+    fp4_quantize,
+)
+from sglang.srt.layers.quantization.modelopt.schemes.modelopt_fp8 import (
+    ModelOptFp8Config,
+    ModelOptFp8KVCacheMethod,
+    ModelOptFp8LinearMethod,
+    ModelOptFp8MoEMethod,
+)
+from sglang.srt.layers.quantization.modelopt.schemes.modelopt_mixed_precision import (
+    ModelOptMixedPrecisionConfig,
+)
+
+__all__ = [
+    "ACTIVATION_SCHEMES",
+    "ACT_STR_TO_TYPE_MAP",
+    "ActivationType",
+    "MOE_NVFP4_DISPATCH",
+    "ModelOptFp4Config",
+    "ModelOptFp4LinearMethod",
+    "ModelOptFp8Config",
+    "ModelOptFp8KVCacheMethod",
+    "ModelOptFp8LinearMethod",
+    "ModelOptFp8MoEMethod",
+    "ModelOptMixedPrecisionConfig",
+    "ModelOptNvFp4FusedMoEMethod",
+    "enable_flashinfer_fp4_gemm",
+    "fp4_gemm",
+    "fp4_quantize",
+]
