+++ python/sglang/srt/layers/quantization/modelopt/schemes/modelopt_fp4.py
@@ -5,7 +5,6 @@ import logging
 from enum import IntEnum
 from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
-import regex as re
 import torch
 from torch.nn.parameter import Parameter
 
@@ -23,32 +22,22 @@ from sglang.srt.layers.moe import (
     get_moe_runner_backend,
 )
 from sglang.srt.layers.moe.cutlass_moe_params import CutlassMoEParams, CutlassMoEType
-from sglang.srt.layers.moe.moe_runner.triton import TritonMoeQuantInfo
 from sglang.srt.layers.moe.utils import should_use_flashinfer_cutlass_moe_fp4_allgather
 from sglang.srt.layers.parameter import ModelWeightParameter, PerTensorScaleParameter
 from sglang.srt.layers.quantization.base_config import (
     FusedMoEMethodBase,
     LinearMethodBase,
-    QuantizationConfig,
-    QuantizeMethodBase,
 )
 from sglang.srt.layers.quantization.fp4_utils import get_fp4_gemm_runner_backend
-from sglang.srt.layers.quantization.fp8_kernel import scaled_fp8_quant
-from sglang.srt.layers.quantization.fp8_utils import (
-    apply_fp8_linear,
-    cutlass_fp8_supported,
-    is_blackwell_supported,
+from sglang.srt.layers.quantization.fp8_utils import is_blackwell_supported
+from sglang.srt.layers.quantization.modelopt.modelopt import ModelOptQuantConfig
+from sglang.srt.layers.quantization.modelopt.utils import (
+    pad_nvfp4_activation_for_cutlass,
+    pad_nvfp4_weight,
+    round_up_to_multiple,
+    slice_nvfp4_output,
 )
-from sglang.srt.layers.quantization.kv_cache import BaseKVCacheMethod
-from sglang.srt.layers.quantization.unquant import UnquantizedLinearMethod
-from sglang.srt.layers.quantization.utils import (
-    convert_to_channelwise,
-    is_layer_skipped,
-    per_tensor_dequantize,
-    requantize_with_max_scale,
-    swizzle_blockscale,
-)
-from sglang.srt.layers.radix_attention import RadixAttention
+from sglang.srt.layers.quantization.utils import swizzle_blockscale
 from sglang.srt.layers.utils import copy_or_rebind_param
 from sglang.srt.utils.common import (
     get_bool_env_var,
@@ -66,7 +55,6 @@ if TYPE_CHECKING:
         CombineInput,
         StandardDispatchOutput,
     )
-    from sglang.srt.models.utils import WeightsMapper
 
 fp4_quantize = None
 try:
@@ -103,10 +91,6 @@ except ImportError:
     class ActivationType(IntEnum):
         Swiglu = 3
         Relu2 = 6
-
-
-# Initialize logger for the module
-logger = logging.getLogger(__name__)
 
 
 def _sglang_fp4_gemm_fake(
@@ -157,103 +141,6 @@ CUTEDSL_MOE_SCALAR_INPUT_SCALE = get_bool_env_var(
     "SGLANG_CUTEDSL_MOE_SCALAR_INPUT_SCALE", "true"
 )
 
-# FP4 GEMM alignment constant - CUTLASS/FlashInfer kernels require dimensions divisible by 32
-FP4_GEMM_ALIGNMENT = 32
-
-
-def round_up_to_multiple(x: int, m: int) -> int:
-    """Round up x to the nearest multiple of m."""
-    return (x + m - 1) // m * m
-
-
-def pad_nvfp4_weight(
-    weight: torch.Tensor,
-    n_alignment: int = FP4_GEMM_ALIGNMENT,
-    k_alignment: int = FP4_GEMM_ALIGNMENT,
-) -> tuple[torch.Tensor, int]:
-    """
-    Pad packed NVFP4 weights to satisfy alignment constraints for FP4 GEMM kernels.
-
-    Different backends have different alignment requirements:
-    - CUTLASS/cuDNN: N % 32 == 0, K % 32 == 0
-    - TRTLLM: N % 128 == 0 (for shuffle_matrix_sf_a), K padding handled separately
-
-    Args:
-        weight: Packed FP4 weight tensor of shape [N, K//2] (2 FP4 values per byte)
-        n_alignment: Required alignment for N dimension (default 32, use 128 for TRTLLM)
-        k_alignment: Required alignment for K dimension (default 32, use 0 to skip)
-
-    Returns:
-        Tuple of (padded_weight, weights_padding_cols) where weights_padding_cols
-        is the number of columns added for K-dimension padding (in bytes).
-    """
-    weight_current_rows = weight.shape[0]  # N dimension
-    weight_current_col_bytes = weight.shape[1]  # K//2 (packed)
-
-    # Calculate padding for N dimension (rows)
-    pad_rows = 0
-    if n_alignment > 0 and weight_current_rows % n_alignment != 0:
-        total_rows = round_up_to_multiple(weight_current_rows, n_alignment)
-        pad_rows = total_rows - weight_current_rows
-
-    # Calculate padding for K dimension (columns)
-    # 2 FP4 items are packed per byte in the input dimension
-    weight_current_col_elements = weight_current_col_bytes * 2
-    pad_cols_bytes = 0
-    if k_alignment > 0 and weight_current_col_elements % k_alignment != 0:
-        total_cols = round_up_to_multiple(weight_current_col_elements, k_alignment)
-        pad_cols = total_cols - weight_current_col_elements
-        # pad_cols is in elements, but padding is in bytes (2 elements per byte)
-        pad_cols_bytes = pad_cols // 2
-
-    # Apply padding in a single operation if needed
-    # For 2D tensor, pad argument is (pad_left, pad_right, pad_top, pad_bottom)
-    if pad_rows > 0 or pad_cols_bytes > 0:
-        weight = torch.nn.functional.pad(
-            weight, (0, pad_cols_bytes, 0, pad_rows)
-        ).contiguous()
-
-    return weight, pad_cols_bytes
-
-
-def pad_nvfp4_activation_for_cutlass(
-    x_fp4: torch.Tensor,
-    weights_padding_cols: int,
-) -> torch.Tensor:
-    """
-    Pad packed FP4 activations to match the K-dimension padding applied to weights.
-
-    Args:
-        x_fp4: Packed FP4 activation tensor
-        weights_padding_cols: Number of padding columns (in bytes) from weight padding
-
-    Returns:
-        Padded activation tensor
-    """
-    if weights_padding_cols > 0:
-        return torch.nn.functional.pad(x_fp4, (0, weights_padding_cols)).contiguous()
-    return x_fp4
-
-
-def slice_nvfp4_output(
-    out: torch.Tensor,
-    output_size: int,
-) -> torch.Tensor:
-    """
-    Slice the output tensor to remove padding in N dimension if weight was padded.
-
-    Args:
-        out: Output tensor from FP4 GEMM
-        output_size: Original output size before padding
-
-    Returns:
-        Sliced output tensor with padding removed
-    """
-    if out.shape[-1] != output_size:
-        return out[..., :output_size].contiguous()
-    return out
-
-
 # TODO make it true by default when the DeepEP PR is merged
 MOE_NVFP4_DISPATCH = envs.SGLANG_MOE_NVFP4_DISPATCH.get()
 # Supported activation schemes for the current configuration
@@ -264,833 +151,7 @@ ACT_STR_TO_TYPE_MAP = {
     "relu2": ActivationType.Relu2,
 }
 
-
-class ModelOptQuantConfig(QuantizationConfig):
-    def __init__(
-        self,
-        kv_cache_quant_algo: Optional[str],
-        exclude_modules: Optional[List[str]],
-        packed_modules_mapping: Optional[Dict[str, List[str]]],
-    ):
-        super().__init__()
-        self.packed_modules_mapping = packed_modules_mapping
-        self.exclude_modules = exclude_modules or []
-        self.kv_cache_quant_algo = kv_cache_quant_algo
-
-    def _get_quant_method(
-        self,
-        layer: torch.nn.Module,
-        prefix: str,
-        *,
-        Linear: type[LinearMethodBase],
-        Moe: type[FusedMoEMethodBase],
-    ) -> Optional[QuantizeMethodBase]:
-        from sglang.srt.layers.linear import LinearBase
-        from sglang.srt.layers.moe.fused_moe_triton import FusedMoE
-
-        if isinstance(layer, LinearBase):
-            if is_layer_skipped(
-                prefix, self.exclude_modules, self.packed_modules_mapping
-            ) or self.is_layer_excluded(prefix):
-                return UnquantizedLinearMethod()
-            return Linear(self)
-        elif self.kv_cache_quant_algo and isinstance(layer, RadixAttention):
-            return ModelOptFp8KVCacheMethod(self)
-        elif isinstance(layer, FusedMoE):
-            # Check if MoE layer should be excluded from quantization
-            # (e.g., MTP layers that have no quantization scales in checkpoint)
-            if self.is_layer_excluded(prefix):
-                # Falls back to default unquantized MoE
-                return None
-            return Moe(self)
-        return None
-
-    @classmethod
-    def get_config_filenames(cls) -> List[str]:
-        return ["hf_quant_config.json"]
-
-    def get_scaled_act_names(self) -> List[str]:
-        return []
-
-    def apply_weight_name_mapper(
-        self, hf_to_sglang_mapper: "WeightsMapper"
-    ):  # noqa: B027
-        # Map excluded module patterns from HF layout to sglang layout.
-        # Ref: HF hf_quant_config.json for nvidia/Kimi-K2.5-NVFP4
-        # https://huggingface.co/nvidia/Kimi-K2.5-NVFP4/blob/main/hf_quant_config.json
-        if self.exclude_modules:
-            mapped = hf_to_sglang_mapper.apply_list(self.exclude_modules)
-            expanded: List[str] = []
-            for name in mapped:
-                expanded.append(name)
-                if name.startswith("language_model."):
-                    expanded.append(name.removeprefix("language_model."))
-            # Preserve order, drop duplicates.
-            self.exclude_modules = list(dict.fromkeys(expanded))
-
-    def is_layer_excluded(self, prefix: str) -> bool:
-        """Check if a layer should be excluded from quantization.
-
-        Handles:
-        - Exact matches (e.g., "lm_head" matching prefix "lm_head")
-        - Glob-style wildcards (e.g., "mtp*" matching "mtp_layers")
-        - Part-by-part matching (split prefix on "." and check each part)
-        - language_model. prefix stripping for vision-language models
-        - Fused module patterns (e.g., "q_a_proj" in "fused_qkv_a_proj_with_mqa")
-        """
-        if not self.exclude_modules:
-            return False
-
-        # Build prefix variants: some models wrap layers under "language_model."
-        prefixes_to_check = [prefix]
-        if prefix.startswith("language_model."):
-            prefixes_to_check.append(prefix.removeprefix("language_model."))
-
-        # Fused module patterns: the exclude list may reference a sub-component
-        # (e.g., "q_a_proj") that is fused into a combined parameter name
-        # (e.g., "fused_qkv_a_proj_with_mqa"). We check if the last segment of
-        # the exclude pattern is a substring of the last segment of the prefix.
-        fused_patterns = {"q_a_proj", "q_b_proj", "kv_a_proj_with_mqa", "kv_b_proj"}
-
-        for pattern in self.exclude_modules:
-            # Convert glob-style wildcard to regex (e.g., "mtp*" -> "mtp.*")
-            regex_str = pattern.replace(".", r"\.").replace("*", r".*")
-
-            for pfx in prefixes_to_check:
-                if re.fullmatch(regex_str, pfx):
-                    return True
-                # Part-by-part check: handles wildcards like "mtp*" matching
-                pfx_parts = pfx.split(".")
-                for part in pfx_parts:
-                    if re.fullmatch(regex_str, part):
-                        return True
-
-            # Check fused patterns: if the last segment of the exclude pattern
-            # is a known fused component, check if it appears in the prefix's
-            # last segment (handles fused_qkv_a_proj_with_mqa containing q_a_proj)
-            pattern_tail = pattern.rsplit(".", maxsplit=1)[-1]
-            if pattern_tail in fused_patterns:
-                for pfx in prefixes_to_check:
-                    if pattern_tail in pfx.rsplit(".", maxsplit=1)[-1]:
-                        return True
-
-        return False
-
-
-class ModelOptFp8Config(ModelOptQuantConfig):
-    """Configuration for ModelOpt FP8 quantization, including serialization and compatibility checks."""
-
-    def __init__(
-        self,
-        is_checkpoint_fp8_serialized: bool = False,
-        kv_cache_quant_method: Optional[str] = None,
-        exclude_modules: Optional[List[str]] = None,
-        packed_modules_mapping: Optional[Dict[str, List[str]]] = None,
-    ) -> None:
-        """
-        Args:
-            is_checkpoint_fp8_serialized (bool): Indicates if the checkpoint uses serialized FP8 format.
-        """
-        super().__init__(kv_cache_quant_method, exclude_modules, packed_modules_mapping)
-        self.is_checkpoint_fp8_serialized = is_checkpoint_fp8_serialized
-        if is_checkpoint_fp8_serialized:
-            logger.warning(
-                "Detected ModelOpt FP8 checkpoint. The format is experimental and subject to change."
-            )
-
-    @classmethod
-    def override_quantization_method(cls, hf_quant_config, user_quant):
-        """Override quantization method based on the model's config."""
-        return cls._modelopt_override_quantization_method(hf_quant_config, user_quant)
-
-    @classmethod
-    def get_name(cls) -> str:
-        return "modelopt_fp8"
-
-    @classmethod
-    def get_supported_act_dtypes(cls) -> List[torch.dtype]:
-        return [torch.bfloat16, torch.half]
-
-    @classmethod
-    def get_min_capability(cls) -> int:
-        return 89  # Minimum hardware capability (e.g., Hopper GPUs).
-
-    @classmethod
-    def from_config(cls, config: Dict[str, Any]) -> ModelOptFp8Config:
-        # Handle two different config formats:
-        # 1. hf_quant_config.json format: {"quantization": {"quant_algo": "FP8", ...}}
-        # 2. config.json quantization_config format: {"quant_algo": "FP8", ...}
-        # In future modelopt will deprecate hf_quant_config.json, and only keep config.json.
-        # For legacy reasons, we keep hf_quant_config.json for now.
-
-        # Initialize variables
-        kv_cache_quant_method = None
-        exclude_modules = None
-
-        # Try flat format first (config.json quantization_config - preferred format)
-        quant_method = config.get("quant_algo")
-        if quant_method is not None:
-            # Flat format (config.json quantization_config)
-            # Derive kv_cache quant from kv_cache_scheme dict
-            kv_cache_scheme = config.get("kv_cache_scheme")
-            if isinstance(kv_cache_scheme, dict):
-                if (
-                    kv_cache_scheme.get("type") == "float"
-                    and kv_cache_scheme.get("num_bits") == 8
-                ):
-                    kv_cache_quant_method = "FP8"
-
-            # Map 'ignore' field to 'exclude_modules'
-            exclude_modules = config.get("ignore")
-        else:
-            # Fall back to nested format (hf_quant_config.json - will be deprecated)
-            try:
-                quantization_section = cls.get_from_keys(config, ["quantization"])
-                quant_method = quantization_section.get("quant_algo")
-                kv_cache_quant_method = quantization_section.get("kv_cache_quant_algo")
-                exclude_modules = quantization_section.get("exclude_modules")
-            except ValueError:
-                raise ValueError(
-                    "Cannot find 'quant_algo' in the model's quantization config. "
-                    "Expected either flat format (config.json) or nested format (hf_quant_config.json)."
-                )
-        if quant_method is None:
-            raise ValueError(
-                "Cannot find 'quant_algo' in the model's quantization config. "
-            )
-        if "FP8" not in quant_method:
-            raise ValueError(
-                "ModelOptFp8Config only supports static FP8 quantization in SGLang. "
-                "For FP4 quantization, use ModelOptFp4Config. "
-                "Check the quantization config for your model's configuration."
-            )
-
-        return cls(
-            is_checkpoint_fp8_serialized=True,
-            kv_cache_quant_method=kv_cache_quant_method,
-            exclude_modules=exclude_modules,
-            packed_modules_mapping=config.get("packed_modules_mapping"),
-        )
-
-    def get_quant_method(
-        self, layer: torch.nn.Module, prefix: str
-    ) -> Optional[QuantizeMethodBase]:
-        return self._get_quant_method(
-            layer, prefix, Linear=ModelOptFp8LinearMethod, Moe=ModelOptFp8MoEMethod
-        )
-
-
-class ModelOptFp8LinearMethod(LinearMethodBase):
-    """Linear method for ModelOpt static FP8 quantization.
-
-    Supports loading FP8 checkpoints with static weight and activation scales.
-    Future support may include dynamic scales.
-
-    **Limitations**:
-    1. Only supports per-tensor quantization due to `torch._scaled_mm` limitations.
-    2. Only supports the `float8_e4m3fn` data type.
-
-    Args:
-        quant_config (ModelOptFp8Config): The ModelOpt quantization configuration.
-    """
-
-    def __init__(self, quant_config: ModelOptFp8Config):
-        super().__init__()
-        self.quant_config = quant_config
-        self.cutlass_fp8_supported = cutlass_fp8_supported()
-
-    def create_weights(
-        self,
-        layer: torch.nn.Module,
-        input_size_per_partition: int,
-        output_partition_sizes: List[int],
-        input_size: Optional[int],
-        output_size: Optional[int],
-        params_dtype: torch.dtype,
-        **extra_weight_attrs,
-    ) -> None:
-        """Creates and registers weights, weight scales, and input scales for FP8 quantization."""
-        output_size_per_partition = sum(output_partition_sizes)
-        weight_loader = extra_weight_attrs.get("weight_loader")
-        weight_dtype = (
-            torch.float8_e4m3fn
-            if self.quant_config.is_checkpoint_fp8_serialized
-            else params_dtype
-        )
-
-        # Set layer attributes
-        layer.logical_widths = output_partition_sizes
-        layer.input_size_per_partition = input_size_per_partition
-        layer.output_size_per_partition = output_size_per_partition
-
-        # Register weight
-        layer.register_parameter(
-            "weight",
-            ModelWeightParameter(
-                data=torch.empty(
-                    output_size_per_partition,
-                    input_size_per_partition,
-                    dtype=weight_dtype,
-                ),
-                input_dim=1,
-                output_dim=0,
-                weight_loader=weight_loader,
-            ),
-        )
-
-        if self.quant_config.is_checkpoint_fp8_serialized:
-            # Register weight and input scales
-            for scale_name in ["weight_scale", "input_scale"]:
-                layer.register_parameter(
-                    scale_name,
-                    PerTensorScaleParameter(
-                        data=torch.full(
-                            (len(output_partition_sizes),),
-                            torch.finfo(torch.float32).min,
-                            dtype=torch.float32,
-                        ),
-                        weight_loader=weight_loader,
-                    ),
-                )
-
-    def process_weights_after_loading(self, layer: torch.nn.Module) -> None:
-        """Requantizes weights after loading using the maximum scale."""
-        max_w_scale, quantized_weight = requantize_with_max_scale(
-            layer.weight, layer.weight_scale, layer.logical_widths
-        )
-        layer.weight = Parameter(quantized_weight.t(), requires_grad=False)
-        # cutlass sgl-kernel only supports per-channel scale
-        if self.cutlass_fp8_supported:
-            max_w_scale = convert_to_channelwise(max_w_scale, layer.logical_widths)
-        layer.weight_scale = Parameter(max_w_scale, requires_grad=False)
-        layer.input_scale = Parameter(layer.input_scale.max(), requires_grad=False)
-
-    def apply(
-        self,
-        layer: torch.nn.Module,
-        x: torch.Tensor,
-        bias: Optional[torch.Tensor] = None,
-    ) -> torch.Tensor:
-        """Applies FP8 linear transformation."""
-        return apply_fp8_linear(
-            input=x,
-            weight=layer.weight,
-            weight_scale=layer.weight_scale,
-            input_scale=layer.input_scale,
-            bias=bias,
-            cutlass_fp8_supported=self.cutlass_fp8_supported,
-        )
-
-
-class ModelOptFp8KVCacheMethod(BaseKVCacheMethod):
-    """
-    Handles loading FP8 kv-cache scaling factors from modelopt quantized checkpoints.
-    """
-
-    def __init__(self, quant_config: ModelOptFp8Config):
-        super().__init__(quant_config)
-
-
-class ModelOptMixedPrecisionConfig(ModelOptQuantConfig):
-    """Configuration for ModelOpt MIXED_PRECISION checkpoints."""
-
-    def __init__(
-        self,
-        kv_cache_quant_algo: Optional[str],
-        exclude_modules: Optional[List[str]],
-        packed_modules_mapping: Optional[Dict[str, List[str]]],
-        quantized_layers: Dict[str, Dict[str, Any]],
-        fp8_config: ModelOptFp8Config,
-        nvfp4_config: "ModelOptFp4Config",
-    ) -> None:
-        super().__init__(kv_cache_quant_algo, exclude_modules, packed_modules_mapping)
-        self.quantized_layers = quantized_layers
-        self.fp8_config = fp8_config
-        self.nvfp4_config = nvfp4_config
-
-    @classmethod
-    def override_quantization_method(cls, hf_quant_config, user_quant):
-        if hf_quant_config is None:
-            return None
-        if hf_quant_config.get("quant_method", "") == "modelopt_mixed":
-            return "modelopt_mixed"
-        return None
-
-    @classmethod
-    def get_name(cls) -> str:
-        return "modelopt_mixed"
-
-    @classmethod
-    def get_supported_act_dtypes(cls) -> List[torch.dtype]:
-        return [torch.bfloat16, torch.half]
-
-    @classmethod
-    def get_min_capability(cls) -> int:
-        return ModelOptFp4Config.get_min_capability()
-
-    @classmethod
-    def from_config(cls, config: Dict[str, Any]) -> "ModelOptMixedPrecisionConfig":
-        kv_cache_quant_algo = None
-        exclude_modules = None
-        quantized_layers = {}
-
-        quant_algo = config.get("quant_algo")
-        if quant_algo is not None:
-            kv_cache_scheme = config.get("kv_cache_scheme")
-            if isinstance(kv_cache_scheme, dict):
-                if (
-                    kv_cache_scheme.get("type") == "float"
-                    and kv_cache_scheme.get("num_bits") == 8
-                ):
-                    kv_cache_quant_algo = "FP8"
-                elif (
-                    kv_cache_scheme.get("type") == "float"
-                    and kv_cache_scheme.get("num_bits") == 4
-                ):
-                    kv_cache_quant_algo = "NVFP4"
-                else:
-                    kv_cache_quant_algo = "auto"
-            exclude_modules = config.get("ignore")
-            quantized_layers = config.get("quantized_layers", {})
-        else:
-            quantization_section = cls.get_from_keys(config, ["quantization"])
-            quant_algo = quantization_section.get("quant_algo")
-            kv_cache_quant_algo = quantization_section.get("kv_cache_quant_algo")
-            exclude_modules = quantization_section.get("exclude_modules")
-            quantized_layers = quantization_section.get("quantized_layers", {})
-
-        if quant_algo != "MIXED_PRECISION":
-            raise ValueError(
-                "ModelOptMixedPrecisionConfig only supports MIXED_PRECISION checkpoints."
-            )
-        if not quantized_layers:
-            raise ValueError(
-                "MIXED_PRECISION quantization requires a non-empty quantized_layers map."
-            )
-
-        group_size = None
-        for layer_info in quantized_layers.values():
-            if layer_info.get("quant_algo", "").upper() == "NVFP4":
-                group_size = layer_info.get("group_size", 16)
-                break
-        if group_size is None:
-            group_size = 16
-
-        packed_modules_mapping = config.get("packed_modules_mapping")
-        fp8_config = ModelOptFp8Config(
-            is_checkpoint_fp8_serialized=True,
-            kv_cache_quant_method=kv_cache_quant_algo,
-            exclude_modules=[],
-            packed_modules_mapping=packed_modules_mapping,
-        )
-        nvfp4_config = ModelOptFp4Config(
-            is_checkpoint_nvfp4_serialized=True,
-            kv_cache_quant_algo=kv_cache_quant_algo,
-            exclude_modules=[],
-            packed_modules_mapping=packed_modules_mapping,
-            group_size=group_size,
-        )
-
-        return cls(
-            kv_cache_quant_algo=kv_cache_quant_algo,
-            exclude_modules=exclude_modules,
-            packed_modules_mapping=packed_modules_mapping,
-            quantized_layers=quantized_layers,
-            fp8_config=fp8_config,
-            nvfp4_config=nvfp4_config,
-        )
-
-    def apply_weight_name_mapper(self, hf_to_sglang_mapper: "WeightsMapper"):
-        super().apply_weight_name_mapper(hf_to_sglang_mapper)
-        if self.quantized_layers:
-            self.quantized_layers = hf_to_sglang_mapper.apply_dict(
-                self.quantized_layers
-            )
-
-    def _resolve_quant_algo(self, prefix: str) -> Optional[str]:
-        if prefix in self.quantized_layers:
-            return self.quantized_layers[prefix]["quant_algo"].upper()
-
-        proj_name = prefix.rsplit(".", 1)[-1]
-        if self.packed_modules_mapping and proj_name in self.packed_modules_mapping:
-            algos = set()
-            base = prefix.rsplit(".", 1)[0]
-            for shard_name in self.packed_modules_mapping[proj_name]:
-                shard_prefix = f"{base}.{shard_name}"
-                if shard_prefix in self.quantized_layers:
-                    algos.add(self.quantized_layers[shard_prefix]["quant_algo"].upper())
-            if len(algos) == 1:
-                return algos.pop()
-            if len(algos) > 1:
-                raise ValueError(
-                    f"Mixed quant_algo within fused layer {prefix}: {algos}. "
-                    "All shards must use the same quantization."
-                )
-
-        prefix_dot = prefix + "."
-        for key, info in self.quantized_layers.items():
-            if key.startswith(prefix_dot):
-                return info["quant_algo"].upper()
-
-        return None
-
-    def get_quant_method(
-        self, layer: torch.nn.Module, prefix: str
-    ) -> Optional[QuantizeMethodBase]:
-        from sglang.srt.layers.linear import LinearBase
-        from sglang.srt.layers.moe.fused_moe_triton import FusedMoE
-
-        quant_algo = self._resolve_quant_algo(prefix)
-
-        if isinstance(layer, LinearBase):
-            if is_layer_skipped(
-                prefix, self.exclude_modules, self.packed_modules_mapping
-            ) or self.is_layer_excluded(prefix):
-                return UnquantizedLinearMethod()
-            if quant_algo == "FP8":
-                return ModelOptFp8LinearMethod(self.fp8_config)
-            if quant_algo == "NVFP4":
-                return ModelOptFp4LinearMethod(self.nvfp4_config)
-            return UnquantizedLinearMethod()
-
-        if self.kv_cache_quant_algo and isinstance(layer, RadixAttention):
-            return ModelOptFp8KVCacheMethod(self.fp8_config)
-
-        if isinstance(layer, FusedMoE):
-            if self.is_layer_excluded(prefix):
-                return None
-            if quant_algo == "FP8":
-                return ModelOptFp8MoEMethod(self.fp8_config)
-            if quant_algo == "NVFP4":
-                return ModelOptNvFp4FusedMoEMethod(self.nvfp4_config)
-            return None
-
-        return None
-
-
-class ModelOptFp8MoEMethod(FusedMoEMethodBase):
-    """MoE method for ModelOpt FP8.
-    Supports loading FP8 checkpoints with static weight scale and activation scale.
-
-    Args:
-        quant_config: The ModelOpt quantization config.
-    """
-
-    def __init__(self, quant_config: ModelOptFp8Config):
-        self.quant_config = quant_config
-        self.cutlass_fp8_supported = cutlass_fp8_supported()
-
-    def create_weights(
-        self,
-        layer: torch.nn.Module,
-        num_experts: int,
-        hidden_size: int,
-        intermediate_size_per_partition: int,
-        params_dtype: torch.dtype,
-        **extra_weight_attrs,
-    ):
-        from sglang.srt.layers.moe.fused_moe_triton import FusedMoeWeightScaleSupported
-
-        # Use FP8 dtype if checkpoint is serialized, otherwise use the default dtype
-        weight_dtype = (
-            torch.float8_e4m3fn
-            if self.quant_config.is_checkpoint_fp8_serialized
-            else params_dtype
-        )
-        weight_loader = extra_weight_attrs.get("weight_loader")
-        num_shards = 2 if layer.moe_runner_config.is_gated else 1
-        intermediate_size = num_shards * intermediate_size_per_partition
-        w13_weight = ModelWeightParameter(
-            data=torch.empty(
-                num_experts,
-                intermediate_size,
-                hidden_size,
-                dtype=weight_dtype,
-            ),
-            input_dim=2,
-            output_dim=1,
-            weight_loader=weight_loader,
-        )
-        layer.register_parameter("w13_weight", w13_weight)
-
-        w2_weight = ModelWeightParameter(
-            data=torch.empty(
-                num_experts,
-                hidden_size,
-                intermediate_size_per_partition,
-                dtype=weight_dtype,
-            ),
-            input_dim=2,
-            output_dim=1,
-            weight_loader=weight_loader,
-        )
-        layer.register_parameter("w2_weight", w2_weight)
-
-        if self.quant_config.is_checkpoint_fp8_serialized:
-            # WEIGHT SCALES - Per-tensor scaling for ModelOpts
-            # Allocate 2 scales for w1 and w3 respectively.
-            # They will be combined to a single scale after weight loading.
-            w13_scale_shape = (num_experts, num_shards)
-            w13_weight_scale = PerTensorScaleParameter(
-                data=torch.full(
-                    w13_scale_shape,
-                    torch.finfo(torch.float32).min,
-                    dtype=torch.float32,
-                ),
-                weight_loader=weight_loader,
-            )
-            w2_weight_scale = PerTensorScaleParameter(
-                data=torch.full(
-                    (num_experts,), torch.finfo(torch.float32).min, dtype=torch.float32
-                ),
-                weight_loader=weight_loader,
-            )
-            layer.register_parameter("w13_weight_scale", w13_weight_scale)
-            layer.register_parameter("w2_weight_scale", w2_weight_scale)
-
-            # Set weight loader attributes for scales
-            extra_weight_attrs.update(
-                {"quant_method": FusedMoeWeightScaleSupported.TENSOR.value}
-            )
-
-            # INPUT SCALES - Per-tensor scaling for ModelOpt
-            w13_input_scale = PerTensorScaleParameter(
-                data=torch.full((num_experts,), 1.0, dtype=torch.float32),
-                weight_loader=weight_loader,
-            )
-            w2_input_scale = PerTensorScaleParameter(
-                data=torch.full((num_experts,), 1.0, dtype=torch.float32),
-                weight_loader=weight_loader,
-            )
-            layer.register_parameter("w13_input_scale", w13_input_scale)
-            layer.register_parameter("w2_input_scale", w2_input_scale)
-
-    def process_weights_after_loading(self, layer: torch.nn.Module) -> None:
-        """Process FP8 MoE weights after loading from serialized checkpoint.
-
-        Only supports pre-quantized checkpoints with FP8 weights and scales.
-        """
-
-        layer.w13_weight = Parameter(layer.w13_weight.data, requires_grad=False)
-        layer.w2_weight = Parameter(layer.w2_weight.data, requires_grad=False)
-
-        # Handle scale parameters
-        if hasattr(layer, "w13_weight_scale") and layer.w13_weight_scale is not None:
-            # Fp8 moe kernel needs single weight scale for w13 per expert.
-            # We take the max of the w1 and w3 scales then dequant and requant each expert.
-            if layer.w13_weight_scale.dim() == 2:  # Shape: (num_experts, 2)
-                # Get the maximum scale across w1 and w3 for each expert
-                max_w13_scales = layer.w13_weight_scale.max(dim=1).values
-
-                # Requantize each expert's weights using the combined scale
-                # w13_weight has shape (num_experts, 2 * intermediate_size_per_partition, hidden_size)
-                # where the first intermediate_size_per_partition rows are w1, the next are w3
-                num_shards = 2 if layer.moe_runner_config.is_gated else 1
-                intermediate_size_per_partition = (
-                    layer.w13_weight.shape[1] // num_shards
-                )
-                for expert_id in range(layer.w13_weight.shape[0]):
-                    start = 0
-                    for shard_id in range(num_shards):  # (w1 and w3) or w13
-                        # Dequantize using the original scale for this shard
-                        dq_weight = per_tensor_dequantize(
-                            layer.w13_weight[expert_id][
-                                start : start + intermediate_size_per_partition, :
-                            ],
-                            layer.w13_weight_scale[expert_id][shard_id],
-                        )
-                        # Requantize using the combined max scale
-                        (
-                            layer.w13_weight[expert_id][
-                                start : start + intermediate_size_per_partition, :
-                            ],
-                            _,
-                        ) = scaled_fp8_quant(dq_weight, max_w13_scales[expert_id])
-
-                        start += intermediate_size_per_partition
-
-                # Update the scale parameter to be per-expert instead of per-shard
-                layer.w13_weight_scale = Parameter(max_w13_scales, requires_grad=False)
-            else:
-                layer.w13_weight_scale = Parameter(
-                    layer.w13_weight_scale.data, requires_grad=False
-                )
-
-        if hasattr(layer, "w2_weight_scale") and layer.w2_weight_scale is not None:
-            layer.w2_weight_scale = Parameter(
-                layer.w2_weight_scale.data, requires_grad=False
-            )
-        if hasattr(layer, "w13_input_scale") and layer.w13_input_scale is not None:
-            layer.w13_input_scale = Parameter(
-                layer.w13_input_scale.max(), requires_grad=False
-            )
-        if hasattr(layer, "w2_input_scale") and layer.w2_input_scale is not None:
-            layer.w2_input_scale = Parameter(
-                layer.w2_input_scale.max(), requires_grad=False
-            )
-
-        # Align FP8 weights to FlashInfer per-tensor kernel layout if enabled
-        if get_moe_runner_backend().is_flashinfer_trtllm():
-            from sglang.srt.layers.moe.moe_runner.flashinfer_trtllm import (
-                align_fp8_moe_weights_for_flashinfer_trtllm,
-            )
-
-            # ModelOpt FP8 stores weights in [Up, Gate] order, so we need to swap
-            align_fp8_moe_weights_for_flashinfer_trtllm(layer, swap_w13_halves=True)
-        elif get_moe_runner_backend().is_flashinfer_cutlass():
-            assert (
-                hasattr(layer, "w13_input_scale") and layer.w13_input_scale is not None
-            )
-            assert hasattr(layer, "w2_input_scale") and layer.w2_input_scale is not None
-            assert (
-                hasattr(layer, "w13_weight_scale")
-                and layer.w13_weight_scale is not None
-            )
-            assert (
-                hasattr(layer, "w2_weight_scale") and layer.w2_weight_scale is not None
-            )
-
-            input_scale = layer.w13_input_scale.to(torch.float32)
-            activation_scale = layer.w2_input_scale.to(torch.float32)
-            w13_weight_scale = layer.w13_weight_scale.to(torch.float32)
-            w2_weight_scale = layer.w2_weight_scale.to(torch.float32)
-
-            layer.fc1_dequant = Parameter(
-                w13_weight_scale * input_scale, requires_grad=False
-            )
-            layer.fc2_quant = Parameter(
-                activation_scale.reciprocal(), requires_grad=False
-            )
-            layer.fc2_dequant = Parameter(
-                activation_scale * w2_weight_scale, requires_grad=False
-            )
-            layer.fc1_input_dequant = Parameter(input_scale, requires_grad=False)
-
-    def create_moe_runner(
-        self, layer: torch.nn.Module, moe_runner_config: MoeRunnerConfig
-    ):
-        self.moe_runner_config = moe_runner_config
-        self.runner = MoeRunner(MoeRunnerBackend.TRITON, moe_runner_config)
-
-    def apply(
-        self,
-        layer: torch.nn.Module,
-        dispatch_output: StandardDispatchOutput,
-    ) -> CombineInput:
-        x = dispatch_output.hidden_states
-        topk_output = dispatch_output.topk_output
-
-        # Fast path: TRT-LLM FP8 per-tensor MoE using BYPASSED TopK routing
-        from sglang.srt.layers.moe.topk import TopKOutputChecker
-
-        if (
-            get_moe_runner_backend().is_flashinfer_trtllm()
-            and TopKOutputChecker.format_is_bypassed(topk_output)
-        ):
-            from sglang.srt.layers.moe.moe_runner.flashinfer_trtllm import (
-                FlashInferTrtllmFp8MoeQuantInfo,
-                fused_experts_none_to_flashinfer_trtllm_fp8,
-            )
-            from sglang.srt.layers.moe.utils import RoutingMethodType
-
-            topk_config = topk_output.topk_config
-
-            # Constraints for ModelOpt FP8 MoE
-            assert (
-                self.moe_runner_config.activation == "silu"
-            ), "Only silu is supported for flashinfer fp8 moe"
-
-            # Enforce Llama4 routing for ModelOpt FP8 MoE for now.
-            # TODO(brayden): support other routing methods
-            assert topk_config.top_k == 1, "ModelOpt FP8 MoE requires top_k==1"
-            assert (
-                not topk_config.num_expert_group
-            ), "ModelOpt FP8 MoE does not support expert grouping"
-            assert (
-                not topk_config.topk_group
-            ), "ModelOpt FP8 MoE does not support grouped top-k"
-
-            quant_info = FlashInferTrtllmFp8MoeQuantInfo(
-                w13_weight=layer.w13_weight,
-                w2_weight=layer.w2_weight,
-                global_num_experts=layer.num_experts,
-                local_expert_offset=layer.moe_ep_rank * layer.num_local_experts,
-                local_num_experts=layer.num_local_experts,
-                intermediate_size=layer.w2_weight.shape[2],
-                routing_method_type=RoutingMethodType.Llama4,
-                block_quant=False,
-                w13_input_scale=layer.w13_input_scale,
-                output1_scales_scalar=layer.output1_scales_scalar,
-                output1_scales_gate_scalar=layer.output1_scales_gate_scalar,
-                output2_scales_scalar=layer.output2_scales_scalar,
-                use_routing_scales_on_input=True,
-            )
-
-            return fused_experts_none_to_flashinfer_trtllm_fp8(
-                dispatch_output, quant_info, self.moe_runner_config
-            )
-
-        if get_moe_runner_backend().is_flashinfer_cutlass():
-            activation = ACT_STR_TO_TYPE_MAP[self.moe_runner_config.activation]
-            assert (
-                (
-                    activation is ActivationType.Relu2
-                    and not self.moe_runner_config.is_gated
-                )
-                or activation is ActivationType.Swiglu
-                and self.moe_runner_config.is_gated
-            ), "Only Relu2 non-gated or Swiglu gated are supported for flashinfer cutlass fp8 moe"
-            topk_weights, topk_ids = topk_output.topk_weights, topk_output.topk_ids
-            x_fp8, _ = scaled_fp8_quant(x, layer.w13_input_scale)
-            output_dtype = x.dtype
-            original_col = x.shape[1]
-            x_sf = None
-
-            with use_symmetric_memory(
-                get_tp_group(), disabled=not is_allocation_symmetric()
-            ):
-                symm_output = torch.empty(
-                    x.shape[0], original_col, dtype=output_dtype, device=x.device
-                )
-            output = flashinfer_cutlass_fused_moe(
-                output=symm_output,
-                input=x_fp8,
-                token_selected_experts=topk_ids.to(torch.int),
-                token_final_scales=topk_weights,
-                fc1_expert_weights=layer.w13_weight,
-                fc2_expert_weights=layer.w2_weight,
-                output_dtype=output_dtype,
-                input_sf=x_sf,
-                quant_scales=[
-                    layer.fc1_dequant,
-                    layer.fc2_quant,
-                    layer.fc2_dequant,
-                    layer.fc1_input_dequant,
-                ],
-                ep_size=layer.moe_ep_size,
-                ep_rank=layer.moe_ep_rank,
-                tp_size=layer.moe_tp_size,
-                tp_rank=layer.moe_tp_rank,
-                tune_max_num_tokens=next_power_of_2(x.shape[0]),
-                activation_type=activation,
-            )[0]
-
-            from sglang.srt.layers.moe.token_dispatcher import StandardCombineInput
-
-            return StandardCombineInput(hidden_states=output)
-
-        quant_info = TritonMoeQuantInfo(
-            w13_weight=layer.w13_weight,
-            w2_weight=layer.w2_weight,
-            use_fp8_w8a8=True,
-            per_channel_quant=False,
-            w13_scale=layer.w13_weight_scale,
-            w2_scale=layer.w2_weight_scale,
-            a13_scale=layer.w13_input_scale,
-            a2_scale=layer.w2_input_scale,
-        )
-
-        return self.runner.run(dispatch_output, quant_info)
+logger = logging.getLogger(__name__)
 
 
 class ModelOptFp4Config(ModelOptQuantConfig):
